Move models to `Model`
